Drop article cache
The problem with the article cache is that it does not evict stale entries unless they are queried again. This means that it effectively adds a memory leak, which, depending on the size of the pages searched, could be substantial.

Feel free to revert this commit to use as a baseline for a cache, just be aware that it has issues that need resolution.